General cleanup and file organization
