+++ BlueprintUI/Sources/BlueprintView/BlueprintView.swift
@@ -363,17 +363,12 @@ public final class BlueprintView: UIView {
         let viewNodes: [(path: ElementPath, node: NativeViewNode)] = renderContext.perform {
             guard let element = element else { return [] }
 
-            if layoutMode == .singlePass {
-                print("--------------------------")
-            }
-
             let result = element.layout(
                 frame: rootFrame,
                 environment: environment,
                 layoutMode: layoutMode
             )
 
-//            print("XXX: \(self.name ?? nil) / \(type(of: element))")
 //            result.dump()
             // TODO: save the LayoutResultNode here for debugging
