+++ BlueprintUI/Sources/Internal/Extensions/CGSize.swift
@@ -8,4 +8,6 @@ extension CGSize {
     static func - (lhs: CGSize, rhs: CGSize) -> CGSize {
         CGSize(width: lhs.width - rhs.width, height: lhs.height - rhs.height)
     }
+
+    static let infinity = Self(width: CGFloat.infinity, height: CGFloat.infinity)
 }
