+++ BlueprintUI/Sources/Internal/SPContentStorage.swift
@@ -1,0 +1,15 @@
+import CoreGraphics
+import Foundation
+
+
+protocol SPContentStorage {
+    func sizeThatFits(
+        proposal: ProposedViewSize,
+        context: MeasureContext
+    ) -> CGSize
+
+    func performSinglePassLayout(
+        proposal: ProposedViewSize,
+        context: SPLayoutContext
+    ) -> [IdentifiedNode]
+}
