+++ BlueprintUI/Sources/Internal/SPL/IdentifiedNode.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+
+typealias IdentifiedNode = (identifier: ElementIdentifier, node: LayoutResultNode)
