+++ BlueprintUI/Sources/Internal/SPL/MeasureContext.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+
+struct MeasureContext {
+
+    var cache: SPCacheNode
+    var environment: Environment
+}
