+++ BlueprintUI/Sources/Internal/SPL/Sizable.swift
@@ -1,0 +1,8 @@
+import CoreGraphics
+import Foundation
+
+
+protocol Sizable {
+
+    func sizeThatFits(proposal: ProposedViewSize, context: MeasureContext) -> CGSize
+}
