+++ BlueprintUI/Sources/Layout/ConstrainedAspectRatio.swift
@@ -169,7 +169,7 @@ public struct ConstrainedAspectRatio: Element {
         }
 
         func placeSubview(in bounds: CGRect, proposal: ProposedViewSize, subview: LayoutSubview) {
-            // NO-OP
+            subview.place(at: bounds.origin, size: bounds.size)
         }
     }
 }
