+++ BlueprintUI/Sources/Layout/Layout.swift
@@ -33,15 +33,6 @@ public protocol Layout: SPLayout {
 
 }
 
-// extension Layout {
-//    public func sizeThatFits(proposal: ProposedViewSize, subviews: LayoutSubviews) -> CGSize {
-//        fatalError("\(type(of: self)) has not implemented single pass layout")
-//    }
-//    public func placeSubviews(in bounds: CGRect, proposal: ProposedViewSize, subviews: LayoutSubviews) {
-//        fatalError("\(type(of: self)) has not implemented single pass layout")
-//    }
-// }
-
 extension Layout where Traits == () {
 
     public static var defaultTraits: () {
