+++ BlueprintUI/Sources/Layout/SPL/LayoutValueKey.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+
+public protocol LayoutValueKey {
+
+    associatedtype Value
+
+    static var defaultValue: Value { get }
+}
+
+
+public enum GenericLayoutValueKey<LayoutType: Layout>: LayoutValueKey {
+
+    public static var defaultValue: LayoutType.Traits {
+        LayoutType.defaultTraits
+    }
+}
