+++ BlueprintUI/Sources/Layout/UserInteractionEnabled.swift
@@ -43,37 +43,6 @@ public struct UserInteractionEnabled: Element {
     }
 }
 
-struct SPAttributes {
-
-    /// Corresponds to `UIView.layer.transform`.
-    public var transform: CATransform3D
-
-    /// Corresponds to `UIView.alpha`.
-    public var alpha: CGFloat
-
-    /// Corresponds to `UIView.isUserInteractionEnabled`.
-    public var isUserInteractionEnabled: Bool
-
-    /// Corresponds to `UIView.isHidden`.
-    public var isHidden: Bool
-
-    internal init(
-        transform: CATransform3D = CATransform3DIdentity,
-        alpha: CGFloat = 1,
-        isUserInteractionEnabled: Bool = true,
-        isHidden: Bool = false
-    ) {
-        self.transform = transform
-        self.alpha = alpha
-        self.isUserInteractionEnabled = isUserInteractionEnabled
-        self.isHidden = isHidden
-    }
-}
-
-protocol AttributeModifier {
-    func attributes() -> SPAttributes
-}
-
 extension Element {
     /// Conditionally enable user interaction of the wrapped element.
     ///
