+++ BlueprintUICommonControls/Sources/Box.swift
@@ -49,7 +49,6 @@ public struct Box: Element {
         if let wrappedElement = wrappedElement {
             return ElementContent(child: wrappedElement)
         } else {
-//            return ElementContent(intrinsicSize: .init(width: 50, height: 50))
             return ElementContent(intrinsicSize: .zero)
         }
     }
